refactoring (api - app, calendar)
