+++ client/src/index.js
@@ -4,8 +4,6 @@ import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
-window.$nameRequest = 'Angreštécie';
-
 ReactDOM.render(
   <React.StrictMode>
     <App />
